No more weird assignments (only weird reduces)
