+++ src/parse-arguments/convertOptionsToParameters/normalizeModifiers/index.ts
@@ -1,3 +1,4 @@
+import { Modifier } from 'types'
 import {
   isCapModifier,
   isDropModifier,
@@ -7,8 +8,7 @@ import {
   isReplaceModifier,
   isRerollModifier,
   isUniqueModifier
-} from 'typeguards'
-import { Modifier } from 'types'
+} from 'utils'
 
 import convertDropOptionsToParameters from './convert-drop-options-to-parameters'
 import convertGreaterLessOptionsToParameters from './convert-greater-less-options-to-parameters'
@@ -18,53 +18,51 @@ import convertRerollOptionsToParameters from './convert-reroll-options-to-parame
 export default function normalizeModifiers(
   modifiers: Array<Modifier<'inclusive' | number>> = []
 ): Array<Modifier<number>> {
-  const newModifiers: Array<Modifier<number>> = []
-
-  modifiers.forEach((modifier) => {
+  return modifiers.map((modifier) => {
     if (isCapModifier(modifier)) {
-      newModifiers.push({
+      return {
         cap: convertGreaterLessOptionsToParameters(modifier.cap)
-      })
+      }
     }
 
     if (isDropModifier(modifier)) {
-      newModifiers.push({ drop: convertDropOptionsToParameters(modifier.drop) })
+      return { drop: convertDropOptionsToParameters(modifier.drop) }
     }
     if (isRerollModifier(modifier)) {
-      newModifiers.push({
+      return {
         reroll: Array.isArray(modifier.reroll)
           ? modifier.reroll.map((option) =>
               convertRerollOptionsToParameters(option)
             )
           : convertRerollOptionsToParameters(modifier.reroll)
-      })
+      }
     }
     if (isReplaceModifier(modifier)) {
-      newModifiers.push({
+      return {
         replace: Array.isArray(modifier.replace)
           ? modifier.replace.map((option) =>
               convertReplaceOptionsToParameters(option)
             )
           : convertReplaceOptionsToParameters(modifier.replace)
-      })
+      }
     }
     if (isUniqueModifier(modifier)) {
-      newModifiers.push({
+      return {
         unique:
           typeof modifier.unique === 'object'
             ? { notUnique: modifier.unique.notUnique.map(Number) }
             : modifier.unique
-      })
+      }
     }
     if (isExplodeModifier(modifier)) {
-      newModifiers.push({ explode: modifier.explode })
+      return { explode: modifier.explode }
     }
     if (isPlusModifier(modifier)) {
-      newModifiers.push({ plus: Number(modifier.plus) })
+      return { plus: Number(modifier.plus) }
     }
     if (isMinusModifier(modifier)) {
-      newModifiers.push({ minus: Number(modifier.minus) })
+      return { minus: Number(modifier.minus) }
     }
+    throw new Error('Unknown Modifier')
   })
-  return newModifiers
 }
