+++ src/utils.ts
@@ -1,5 +1,22 @@
-import { allPatterns } from 'patterns'
-import { Match, Randomizer } from 'types'
+import { allPatterns, coreNotationPattern } from 'patterns'
+import {
+  CapModifier,
+  DetailedType,
+  DiceNotation,
+  DieType,
+  DropModifier,
+  ExplodeModifier,
+  Match,
+  MinusModifier,
+  Modifier,
+  NumberStringArgument,
+  PlusModifier,
+  Randomizer,
+  RandsumOptions,
+  ReplaceModifier,
+  RerollModifier,
+  UniqueModifier
+} from 'types'
 
 export const completeRollPattern = new RegExp(
   `${allPatterns.map((pattern) => pattern.source).join('|')}`,
@@ -52,3 +69,63 @@ export function rollOneFactory(sides: number, randomizer: Randomizer) {
     return randomizer(sides)
   }
 }
+
+export function isDiceNotation(argument: unknown): argument is DiceNotation {
+  return !!coreNotationPattern.test(String(argument))
+}
+
+export function isRandsumOptions(
+  argument: unknown
+): argument is RandsumOptions<DieType, DetailedType> {
+  return (
+    typeof argument === 'object' &&
+    (argument as RandsumOptions<DieType, DetailedType>).sides !== undefined
+  )
+}
+
+function isModifierType<T extends Modifier<NumberStringArgument>>(
+  argument: Modifier<NumberStringArgument>,
+  key: keyof T
+): argument is T {
+  return (argument as T)[key] !== undefined
+}
+
+export const isRerollModifier = (
+  modifier: Modifier<NumberStringArgument>
+): modifier is RerollModifier<NumberStringArgument> =>
+  isModifierType<RerollModifier<NumberStringArgument>>(modifier, 'reroll')
+
+export const isUniqueModifier = (
+  modifier: Modifier<NumberStringArgument>
+): modifier is UniqueModifier<NumberStringArgument> =>
+  isModifierType<UniqueModifier<NumberStringArgument>>(modifier, 'unique')
+
+export const isReplaceModifier = (
+  modifier: Modifier<NumberStringArgument>
+): modifier is ReplaceModifier<NumberStringArgument> =>
+  isModifierType<ReplaceModifier<NumberStringArgument>>(modifier, 'replace')
+
+export const isCapModifier = (
+  modifier: Modifier<NumberStringArgument>
+): modifier is CapModifier<NumberStringArgument> =>
+  isModifierType<CapModifier<NumberStringArgument>>(modifier, 'cap')
+
+export const isDropModifier = (
+  modifier: Modifier<NumberStringArgument>
+): modifier is DropModifier<NumberStringArgument> =>
+  isModifierType<DropModifier<NumberStringArgument>>(modifier, 'drop')
+
+export const isExplodeModifier = (
+  modifier: Modifier<NumberStringArgument>
+): modifier is ExplodeModifier =>
+  isModifierType<ExplodeModifier>(modifier, 'explode')
+
+export const isPlusModifier = (
+  modifier: Modifier<NumberStringArgument>
+): modifier is PlusModifier<NumberStringArgument> =>
+  isModifierType<PlusModifier<NumberStringArgument>>(modifier, 'plus')
+
+export const isMinusModifier = <T extends NumberStringArgument>(
+  modifier: Modifier<T>
+): modifier is MinusModifier<T> =>
+  isModifierType<MinusModifier<T>>(modifier, 'minus')
